bpf: move egress gateway logic to lib/egress_gateway.h
