+++ bpf/lib/egress_gateway.h
@@ -1,0 +1,38 @@
+/* SPDX-License-Identifier: GPL-2.0 */
+/* Copyright (C) 2021 Authors of Cilium */
+
+#ifndef __EGRESS_GATEWAY_H_
+#define __EGRESS_GATEWAY_H_
+
+#ifdef ENABLE_EGRESS_GATEWAY
+
+#include <bpf/compiler.h>
+#include <bpf/ctx/ctx.h>
+
+#include <linux/ip.h>
+
+#include "maps.h"
+
+/* EGRESS_STATIC_PREFIX represents the size in bits of the static prefix part of
+ * an egress policy key (i.e. the source IP).
+ */
+#define EGRESS_STATIC_PREFIX (sizeof(__be32) * 8)
+#define EGRESS_PREFIX_LEN(PREFIX) (EGRESS_STATIC_PREFIX + (PREFIX))
+#define EGRESS_IPV4_PREFIX EGRESS_PREFIX_LEN(32)
+
+static __always_inline __maybe_unused struct egress_info *
+egress_lookup4(struct bpf_elf_map *map, __be32 sip, __be32 dip)
+{
+	struct egress_key key = {
+		.lpm_key = { EGRESS_IPV4_PREFIX, {} },
+		.sip = sip,
+		.dip = dip,
+	};
+	return map_lookup_elem(map, &key);
+}
+
+#define lookup_ip4_egress_endpoint(sip, dip) \
+	egress_lookup4(&EGRESS_MAP, sip, dip)
+
+#endif /* ENABLE_EGRESS_GATEWAY */
+#endif /* __EGRESS_GATEWAY_H_ */
